Update section checking logic
Updates the logic when checking if a changelog entry is within a
section. The new logic will look through the full changelog after
the changes are applied finding each entry and looking backwards
through the lines looking for a section header. It also looks for
version lines to make sure that a version header is not found before
the section header as that would mean the entry isn't actually within
that section.